requirements.txt
boto3==1.34.47
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.0
PyAthena==3.3.0
python-dotenv==1.0.1
scipy==1.12.0
seaborn==0.13.2
